remove unnecessary float lookups from numpy module

+++ astropy/io/ascii/tests/test_types.py
@@ -11,7 +11,7 @@ from .common import assert_equal
 
 
 def test_types_from_dat():
-    converters = {'a': [ascii.convert_numpy(np.float)],
+    converters = {'a': [ascii.convert_numpy(float)],
                   'e': [ascii.convert_numpy(np.str)]}
 
     dat = ascii.read(['a b c d e', '1 1 cat 2.1 4.2'],
